@@ -1,57 +1,40 @@
 <script lang="ts">
-	import { createEventDispatcher } from 'svelte';
 	import { QueryClient, QueryClientProvider, createQuery } from '@tanstack/svelte-query';
 	import { dbStatus } from '$lib/stores/dbStatus';
 	import { get } from 'svelte/store';
 
-	export let cuid: string;
-	export let title: string;
-	export let node_id: string;
-	export let status: string;
-	export let last_updated: string;
-	export let schemas: string[];
-
-	let errorMessage: string = '';
-
-	let statusColor: string;
-	const dispatch = createEventDispatcher();
-
 	const queryClient = new QueryClient();
 
-	let isDbOnline: boolean = get(dbStatus);
+	interface Props {
+		cuid: string;
+		title: string;
+		node_id: string;
+		status: string;
+		last_updated: string;
+		schemas: string[];
+		profileErrorOccurred: (error: string | null) => void;
+		profileUpdated: () => void;
+		profileModify: (cuid: string) => Promise<void>;
+	}
+
+	let {
+		cuid,
+		title,
+		node_id,
+		status,
+		last_updated,
+		schemas,
+		profileErrorOccurred,
+		profileUpdated,
+		profileModify
+	}: Props = $props();
+
+	let errorMessage: string = $state('');
+	let statusColor: string = $state('variant-filled-success');
+	let isDbOnline: boolean = $state(get(dbStatus));
 
 	// Subscribe to dbStatus changes
-	dbStatus.subscribe((value) => {
-		isDbOnline = value;
-	});
-
-	// Function to fetch status
-	async function fetchStatus(): Promise<string> {
-		if (node_id === '') return 'unknown';
-		if (['posted', 'deleted', 'validation_failed', 'post_failed'].includes(status)) {
-			return status;
-		}
-
-		const response = await fetch(`/profile-generator/index/${node_id}`);
-		const data = await response.json();
-		if (response.ok) {
-			errorMessage = '';
-			dispatch('profileErrorOccurred', null);
-			return data.status ?? 'unknown';
-		} else {
-			if (response.status === 404) {
-				return 'Not found';
-			}
-
-			console.error('Failed to fetch status:', data.error || response.statusText);
-			errorMessage = data.error
-				? data.error
-				: `Unknown error occurred. HTTP Status: ${response.status}. Please try again in a few minutes.`;
-			dispatch('profileErrorOccurred', errorMessage);
-
-			return 'unknown';
-		}
-	}
+	dbStatus.subscribe((value) => (isDbOnline = value));
 
 	// Use svelte-query to fetch status
 	const statusQuery = createQuery({
@@ -60,64 +43,92 @@
 		refetchInterval: 5000
 	});
 
-	$: {
+	$effect(() => {
 		status = $statusQuery.data ?? status;
+		statusColor = getStatusColor(status);
+	});
 
-		switch (status) {
-			case 'posted':
-				statusColor = 'variant-filled-success';
-				break;
-			case 'received':
-			case 'validated':
-				statusColor = 'variant-filled-warning';
-				break;
-			default:
-				statusColor = 'variant-filled-error';
-				break;
+	// Function to fetch status
+	async function fetchStatus(): Promise<string> {
+		if (!node_id) return 'unknown';
+		if (['posted', 'deleted', 'validation_failed', 'post_failed'].includes(status)) {
+			return status;
+		}
+
+		try {
+			const response = await fetch(`/profile-generator/index/${node_id}`);
+			const data = await response.json();
+			if (response.ok) {
+				errorMessage = '';
+				profileErrorOccurred(null);
+				return data.status ?? 'unknown';
+			} else {
+				handleFetchError(response, data);
+				return 'unknown';
+			}
+		} catch (error) {
+			console.error('Error fetching status:', error);
+			return 'unknown';
 		}
 	}
 
-	function handleDelete() {
+	// Handle fetch error
+	function handleFetchError(response: Response, data: { error?: string }) {
+		if (response.status === 404) {
+			return 'Not found';
+		}
+		console.error('Failed to fetch status:', data.error || response.statusText);
+		errorMessage =
+			data.error ||
+			`Unknown error occurred. HTTP Status: ${response.status}. Please try again in a few minutes.`;
+		profileErrorOccurred(errorMessage);
+	}
+
+	// Get status color
+	function getStatusColor(status: string): string {
+		switch (status) {
+			case 'posted':
+				return 'variant-filled-success';
+			case 'received':
+			case 'validated':
+				return 'variant-filled-warning';
+			default:
+				return 'variant-filled-error';
+		}
+	}
+
+	function handleDelete(): void {
 		if (confirm('Are you sure you want to delete this profile?')) {
 			deleteProfile();
 		}
 	}
 
-	async function deleteProfile() {
+	async function deleteProfile(): Promise<void> {
 		try {
-			// Delete profile
-			const response = await fetch(`/profile-generator/${cuid}`, {
-				method: 'DELETE',
-				headers: { 'Content-Type': 'application/json' }
-			});
-
-			const result = await response.json();
-			if (result.success) {
-				console.log('Profile deleted successfully');
-			} else {
-				console.error('Failed to delete profile:', result.error);
-			}
-
-			// Delete index if node_id exists
+			await performDelete(`/profile-generator/${cuid}`);
 			if (node_id) {
-				const indexResponse = await fetch(`/profile-generator/index/${node_id}`, {
-					method: 'DELETE',
-					headers: { 'Content-Type': 'application/json' }
-				});
-
-				if (!indexResponse.ok) {
-					console.log('Failed to delete index:', (await indexResponse.json()).error);
-				}
+				await performDelete(`/profile-generator/index/${node_id}`);
 			}
-
-			dispatch('profileUpdated');
+			profileUpdated();
 		} catch (error) {
 			console.error('Error deleting profile:', error);
 		}
 	}
 
-	function handleModify() {
-		dispatch('profileModify', { cuid });
+	// Perform delete operation
+	async function performDelete(url: string): Promise<void> {
+		const response = await fetch(url, {
+			method: 'DELETE',
+			headers: { 'Content-Type': 'application/json' }
+		});
+		const result = await response.json();
+		if (!result.success) {
+			console.error('Failed to delete:', result.error);
+		}
+	}
+
+	function handleModify(): void {
+		profileModify(cuid);
 	}
 </script>
 
@@ -137,12 +148,12 @@
 		</div>
 		<div class="flex justify-around mt-4 md:mt-8">
 			<button
-				on:click={handleModify}
+				onclick={handleModify}
 				class="btn font-semibold md:btn-lg variant-filled-primary"
 				disabled={!!errorMessage || !isDbOnline}>Modify</button
 			>
 			<button
-				on:click={handleDelete}
+				onclick={handleDelete}
 				class="btn font-semibold md:btn-lg variant-filled-secondary"
 				disabled={!!errorMessage || !isDbOnline}>Delete</button
 			>
