@@ -1,7 +1,7 @@
-import type { Handle } from '@sveltejs/kit';
-import { parse } from 'cookie';
 import { closeDatabaseConnection, connectToDatabase } from '$lib/db';
 import { dbStatus } from '$lib/stores/dbStatus';
+import type { Handle } from '@sveltejs/kit';
+import { parse } from 'cookie';
 
 export const handle: Handle = async ({ event, resolve }) => {
 	const cookieHeader = event.request.headers.get('cookie') || '';
@@ -38,8 +38,6 @@ export const handle: Handle = async ({ event, resolve }) => {
 			console.error(`MongoDB connection failed: ${error}`);
 			dbStatus.set(false);
 			event.locals.isAuthenticated = true;
-		} finally {
-			await closeDatabaseConnection();
 		}
 	} else {
 		event.locals.isAuthenticated = false;
@@ -47,3 +45,8 @@ export const handle: Handle = async ({ event, resolve }) => {
 
 	return resolve(event);
 };
+
+process.on('sveltekit:shutdown', async () => {
+	await closeDatabaseConnection();
+	process.exit(0);
+});
