@@ -1,4 +1,6 @@
-export const load = ({ locals }) => {
+import type { LayoutServerLoad } from './$types';
+
+export const load: LayoutServerLoad = ({ locals }) => {
 	return {
 		isAuthenticated: locals.isAuthenticated
 	};
