@@ -1,4 +1,4 @@
-import { closeDatabaseConnection, connectToDatabase } from '$lib/db';
+import { connectToDatabase } from '$lib/db';
 import type { RequestHandler } from '@sveltejs/kit';
 
 export const GET: RequestHandler = async () => {
@@ -9,7 +9,5 @@ export const GET: RequestHandler = async () => {
 	} catch (error) {
 		console.error('Error checking DB status:', error);
 		return new Response(JSON.stringify({ status: 'down' }), { status: 500 });
-	} finally {
-		await closeDatabaseConnection();
 	}
 };
