@@ -1,7 +1,7 @@
-import { json, type RequestHandler } from '@sveltejs/kit';
-import { closeDatabaseConnection, connectToDatabase } from '$lib/db';
-import { serialize } from 'cookie';
 import { PUBLIC_ENV } from '$env/static/public';
+import { connectToDatabase } from '$lib/db';
+import { json, type RequestHandler } from '@sveltejs/kit';
+import { serialize } from 'cookie';
 
 export const POST: RequestHandler = async ({ cookies, locals }) => {
 	try {
@@ -10,7 +10,6 @@ export const POST: RequestHandler = async ({ cookies, locals }) => {
 		if (sessionToken) {
 			const db = await connectToDatabase();
 			await db.collection('sessions').deleteOne({ session_token: sessionToken });
-			await closeDatabaseConnection();
 		}
 
 		const cookieHeader = serialize('murmurations_tools_session', '', {
