@@ -1,5 +1,5 @@
+import { connectToDatabase } from '$lib/db';
 import { json, type RequestHandler } from '@sveltejs/kit';
-import { connectToDatabase, closeDatabaseConnection } from '$lib/db';
 
 export const GET: RequestHandler = async ({ params }) => {
 	const { profile_cuid } = params;
@@ -23,7 +23,5 @@ export const GET: RequestHandler = async ({ params }) => {
 	} catch (error) {
 		console.error('Error fetching profile:', error);
 		return json({ success: false, error: 'Internal server error' }, { status: 500 });
-	} finally {
-		await closeDatabaseConnection();
 	}
 };
